save each frame's timestamp

+++ src/main.cpp
@@ -1,10 +1,10 @@
 #include "data_recorder.hpp"
 
 int main(int argc, char** argv) {
-    int videoLength = 3; // seconds
+    double videoLength = 3; // seconds
     if (argc == 2) {
         try {
-            videoLength = std::stoi(argv[1]);
+            videoLength = std::stod(argv[1]);
         } catch (std::invalid_argument &e) {
             std::cout << "videoLength set to default value: " << videoLength << std::endl;
         }
